+++ examples/dbclient/jdbc/pom.xml
@@ -143,6 +143,10 @@
             <scope>test</scope>
         </dependency>
         <dependency>
+            <groupId>org.testcontainers</groupId>
+            <artifactId>oracle-xe</artifactId>
+        </dependency>
+        <dependency>
             <groupId>com.mysql</groupId>
             <artifactId>mysql-connector-j</artifactId>
             <scope>test</scope>
@@ -194,6 +198,11 @@
                             <goal>integration-test</goal>
                             <goal>verify</goal>
                         </goals>
+                        <configuration>
+                            <systemProperties>
+                                <oracle.jdbc.timezoneAsRegion>false</oracle.jdbc.timezoneAsRegion>
+                            </systemProperties>
+                        </configuration>
                     </execution>
                 </executions>
             </plugin>
