+++ examples/dbclient/jdbc/src/test/java/io/helidon/examples/dbclient/jdbc/PokemonServiceMySQLIT.java
@@ -27,6 +27,8 @@ import org.testcontainers.containers.MySQLContainer;
 import org.testcontainers.junit.jupiter.Container;
 import org.testcontainers.junit.jupiter.Testcontainers;
 
+import static io.helidon.config.ConfigSources.classpath;
+
 @Testcontainers(disabledWithoutDocker = true)
 class PokemonServiceMySQLIT extends AbstractPokemonServiceTest {
 
@@ -40,12 +42,8 @@ class PokemonServiceMySQLIT extends AbstractPokemonServiceTest {
     @BeforeAll
     static void start() {
         Config.global(Config.builder()
-                .addSource(ConfigSources.create(
-                        Map.of("db.connection.url", container.getJdbcUrl(),
-                                "db.connection.username", "user",
-                                "db.connection.password", "password",
-                                "db.connection.poolName", "mysql")))
-                .addSource(ConfigSources.create(Config.create()))
+                .addSource(ConfigSources.create(Map.of("db.connection.url", container.getJdbcUrl())))
+                .addSource(classpath("application-mysql-test.yaml"))
                 .build());
         beforeAll();
     }
