+++ examples/dbclient/jdbc/src/test/java/io/helidon/examples/dbclient/jdbc/PokemonServiceOracleIT.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023, 2024 Oracle and/or its affiliates.
+ * Copyright (c) 2024 Oracle and/or its affiliates.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.helidon.examples.dbclient.pokemons;
+package io.helidon.examples.dbclient.jdbc;
 
 import java.util.Map;
 
@@ -22,7 +22,7 @@ import io.helidon.config.ConfigSources;
 
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
-import org.testcontainers.containers.GenericContainer;
+import org.testcontainers.containers.OracleContainer;
 import org.testcontainers.containers.wait.strategy.Wait;
 import org.testcontainers.junit.jupiter.Container;
 import org.testcontainers.junit.jupiter.Testcontainers;
@@ -30,24 +30,24 @@ import org.testcontainers.utility.DockerImageName;
 
 import static io.helidon.config.ConfigSources.classpath;
 
-/**
- * Tests {@link io.helidon.examples.dbclient.pokemons.PokemonService}.
- */
 @Testcontainers(disabledWithoutDocker = true)
-class PokemonServiceH2IT extends AbstractPokemonServiceTest {
-    private static final DockerImageName H2_IMAGE = DockerImageName.parse("nemerosa/h2");
+public class PokemonServiceOracleIT extends AbstractPokemonServiceTest {
+
+    private static final DockerImageName image = DockerImageName.parse("wnameless/oracle-xe-11g-r2")
+            .asCompatibleSubstituteFor("gvenzl/oracle-xe");
 
     @Container
-    static GenericContainer<?> container = new GenericContainer<>(H2_IMAGE)
-            .withExposedPorts(9082)
-            .waitingFor(Wait.forLogMessage("(.*)Web Console server running at(.*)", 1));
+    static OracleContainer container = new OracleContainer(image)
+            .withExposedPorts(1521, 8080)
+            .withDatabaseName("XE")
+            .usingSid()
+            .waitingFor(Wait.forListeningPorts(1521, 8080));
 
     @BeforeAll
     static void start() {
-        String url = String.format("jdbc:h2:tcp://localhost:%s/~./test", container.getMappedPort(9082));
         Config.global(Config.builder()
-                .addSource(ConfigSources.create(Map.of("db.connection.url", url)))
-                .addSource(classpath("application-h2-test.yaml"))
+                .addSource(ConfigSources.create(Map.of("db.connection.url", container.getJdbcUrl())))
+                .addSource(classpath("application-oracle-test.yaml"))
                 .build());
         beforeAll();
     }
