+++ examples/dbclient/pokemons/src/test/resources/application-mongo-test.yaml
@@ -1,0 +1,113 @@
+#
+# Copyright (c) 2024 Oracle and/or its affiliates.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+server:
+  port: 8080
+  host: 0.0.0.0
+
+tracing:
+  service: mongo-db
+
+db:
+  source: "mongoDb"
+  init-schema: false
+  # Transactions are not supported
+  init-data: false
+  services:
+    tracing:
+      - enabled: true
+  health-check:
+    type: "query"
+    statementName: "health-check"
+  statements:
+    # Health check statement. HealthCheck statement type must be a query.
+    health-check: '{
+            "operation": "command",
+            "query": { ping: 1 }
+        }'
+    ## Create database schema
+    # Select all types
+    select-all-types: '{
+            "collection": "types",
+            "operation": "query",
+            "projection": { id: 1, name: 1, _id: 0 },
+            "query": {}
+        }'
+    # Select all pokemons without type information
+    select-all-pokemons: '{
+            "collection": "pokemons",
+            "operation": "query",
+            "projection": { id: 1, name: 1, id_type: 1, _id: 0 },
+            "query": {}
+        }'
+    # Select pokemon by id
+    select-pokemon-by-id: '{
+            "collection": "pokemons",
+            "operation": "query",
+            "projection": { id: 1, name: 1, id_type: 1, _id: 0 },
+            "query": { id: $id }
+        }'
+    # Select pokemon by name
+    select-pokemon-by-name: '{
+            "collection": "pokemons",
+            "operation": "query",
+            "projection": { id: 1, name: 1, id_type: 1, _id: 0 },
+            "query": { name: ? }
+        }'
+    # Insert records into database
+    insert-type: '{
+            "collection": "types",
+            "operation": "insert",
+            "value": {
+                "id": ?,
+                "name": ?
+            }
+        }'
+    insert-pokemon: '{
+            "collection": "pokemons",
+            "operation": "insert",
+            "value": {
+                "id": ?,
+                "name": ?,
+                "id_type": ?
+            }
+        }'
+    # Update name of pokemon specified by id
+    update-pokemon-by-id: '{
+            "collection": "pokemons",
+            "operation": "update",
+            "value":{ $set: { "name": $name, "id_type": $idType } },
+            "query": { id: $id }
+        }'
+    # Delete pokemon by id
+    delete-pokemon-by-id: '{
+            "collection": "pokemons",
+            "operation": "delete",
+            "query": { id: $id }
+        }'
+    # Delete all types
+    delete-all-types: '{
+            "collection": "types",
+            "operation": "delete",
+            "query": { }
+        }'
+    # Delete all pokemons
+    delete-all-pokemons: '{
+            "collection": "pokemons",
+            "operation": "delete",
+            "query": { }
+        }'
+
